+++ R/casechange-functions.R
@@ -16,15 +16,21 @@ tokens_tolower <- function(x, keep_acronyms = FALSE, ...) {
 
 #' @noRd
 #' @export
-tokens_tolower.tokenizedTexts <- function(x, keep_acronyms = FALSE, ...) {
-    as.tokenizedTexts(tokens_tolower(as.tokens(x), keep_acronyms = keep_acronyms, ...))
+tokens_tolower.tokens <- function(x, keep_acronyms = FALSE, ...) {
+    types(x) <- lowercase_types(types(x), keep_acronyms)
+    tokens_recompile(x)
 }
 
 #' @noRd
-#' @export
-tokens_tolower.tokens <- function(x, keep_acronyms = FALSE, ...) {
-    types(x) <- char_tolower(types(x), keep_acronyms = keep_acronyms, ...)
-    tokens_recompile(x)
+#' @keywords internal
+lowercase_types <- function(type, keep_acronyms) {
+    if (keep_acronyms) {
+        is_acronyms <- stri_detect_regex(type, "^\\p{Uppercase_Letter}(\\p{Uppercase_Letter}|\\d)+$")
+    } else {
+        is_acronyms <- rep(FALSE, length(type))
+    }
+    type[!is_acronyms] <- stri_trans_tolower(type[!is_acronyms])
+    return(type)
 }
 
 
@@ -33,12 +39,6 @@ tokens_tolower.tokens <- function(x, keep_acronyms = FALSE, ...) {
 #' @export
 tokens_toupper <- function(x, ...) {
     UseMethod("tokens_toupper")
-}
-
-#' @noRd
-#' @export
-tokens_toupper.tokenizedTexts <- function(x, ...) {
-    as.tokenizedTexts(tokens_toupper(as.tokens(x), ...))
 }
     
 #' @noRd
@@ -83,18 +83,21 @@ char_tolower <- function(x, keep_acronyms = FALSE, ...) {
 #' @noRd
 #' @export
 char_tolower.character <- function(x, keep_acronyms = FALSE, ...) {
-    savedNames <- names(x)
-    if (keep_acronyms)
-        x <- stri_replace_all_regex(x, "\\b(\\p{Uppercase_Letter}{2,})\\b",  "_$1_", ...)
-    x <- stri_trans_tolower(x, ...)
+    name <- names(x)
     if (keep_acronyms) {
-        m1 <- unique(unlist(stri_extract_all_regex(x, "\\b_\\p{Lowercase_Letter}+_\\b", omit_no_match = TRUE, ...)))
-        if (length(m1) > 0) {
-            m2 <- stri_replace_all_fixed(stri_trans_toupper(m1, ...), "_", "", ...)
-            x <- vapply(x, function(s) stri_replace_all_regex(s, m1,  m2, vectorize_all = FALSE, ...), character(1))
+        match <- stri_extract_all_regex(x, "\\b(\\p{Uppercase_Letter}(\\p{Uppercase_Letter}|\\d)+)\\b")
+        for (i in which(lengths(match) > 0)) {
+            m <- unique(match[[i]])
+            x[i] <- stri_replace_all_regex(x[i], paste0('\\b', m, '\\b'), 
+                                                 paste0('\uE000', m, '\uE001'), vectorize_all = FALSE)
+            x[i] <- stri_trans_tolower(x[i])
+            x[i] <- stri_replace_all_regex(x[i], paste0('\uE000', stri_trans_tolower(m), '\uE001'), 
+                                                 m, vectorize_all = FALSE)
         }
+    } else {
+        x <- stri_trans_tolower(x)
     }
-    names(x) <- savedNames
+    names(x) <- name
     return(x)
 }
 
@@ -107,9 +110,9 @@ char_toupper <- function(x, ...) {
 #' @noRd
 #' @export 
 char_toupper.character <- function(x, ...) {
-    savedNames <- names(x)
+    name <- names(x)
     x <- stri_trans_toupper(x, ...)
-    names(x) <- savedNames
+    names(x) <- name
     return(x)
 }
 
@@ -136,7 +139,7 @@ dfm_tolower <- function(x, keep_acronyms = FALSE, ...) {
 #' @export
 dfm_tolower.dfm <- function(x, keep_acronyms = FALSE, ...) {
     x <- as.dfm(x)
-    colnames(x) <- char_tolower(featnames(x), keep_acronyms = keep_acronyms, ...)
+    colnames(x) <- lowercase_types(featnames(x), keep_acronyms)
     dfm_compress(x, margin = "features")
 }
 
@@ -175,8 +178,7 @@ fcm_tolower <- function(x, keep_acronyms = FALSE, ...) {
 #' @noRd
 #' @export
 fcm_tolower.fcm <- function(x, keep_acronyms = FALSE, ...) {
-    colnames(x) <- rownames(x) <- 
-        char_tolower(colnames(x), keep_acronyms = keep_acronyms, ...)
+    colnames(x) <- rownames(x) <- lowercase_types(featnames(x), keep_acronyms)
     fcm_compress(x)
 }