+++ R/dfm_select.R
@@ -1,9 +1,10 @@
 #' select features from a dfm or fcm
 #' 
-#' This function selects or discards features from a \link{dfm} or \link{fcm}, 
+#' This function selects or removes features from a \link{dfm} or \link{fcm}, 
 #' based on feature name matches with \code{pattern}.  The most common usages 
 #' are to eliminate features from a dfm already constructed, such as stopwords, 
 #' or to select only terms of interest from a dictionary.
+#' 
 #' @param x the \link{dfm} or \link{fcm} object whose features will be selected
 #' @inheritParams pattern
 #' @param selection whether to \code{keep} or \code{remove} the features
@@ -27,6 +28,10 @@
 #' @details \code{dfm_remove} and \code{fcm_remove} are simply a convenience
 #'   wrappers to calling \code{dfm_select} and \code{fcm_select} with 
 #'   \code{selection = "remove"}.
+#'   
+#'   \code{dfm_keep} and \code{fcm_keep} are simply a convenience
+#'   wrappers to calling \code{dfm_select} and \code{fcm_select} with 
+#'   \code{selection = "keep"}.
 #' @note This function selects features based on their labels.  To select 
 #'   features based on the values of the document-feature matrix, use 
 #'   \code{\link{dfm_trim}}.
@@ -76,7 +81,7 @@ dfm_select <- function(x, pattern = NULL,
                        selection = c("keep", "remove"), 
                        valuetype = c("glob", "regex", "fixed"),
                        case_insensitive = TRUE,
-                       min_nchar = 1L, max_nchar = 63L,
+                       min_nchar = 1L, max_nchar = 79L,
                        verbose = quanteda_options("verbose"), ...) {
     UseMethod("dfm_select")
 }
@@ -88,7 +93,7 @@ dfm_select.dfm <-  function(x, pattern = NULL,
                             selection = c("keep", "remove"), 
                             valuetype = c("glob", "regex", "fixed"),
                             case_insensitive = TRUE,
-                            min_nchar = 1L, max_nchar = 63L,
+                            min_nchar = 1L, max_nchar = 79L,
                             verbose = quanteda_options("verbose"), ...) {
     
     x <- as.dfm(x)
@@ -165,15 +170,21 @@ dfm_select.dfm <-  function(x, pattern = NULL,
 #'               verbose = FALSE)
 #' tmpdfm
 #' dfm_remove(tmpdfm, stopwords("english"))
-dfm_remove <- function(x, pattern = NULL, ...) {
-    UseMethod("dfm_remove")
+dfm_remove <- function(x, ...) {
+    if ("selection" %in% names(list(...))) {
+        stop("dfm_remove cannot include selection argument")
+    }
+    dfm_select(x, ..., selection = "remove")
 }
 
-#' @noRd
+
+#' @rdname dfm_select
 #' @export
-dfm_remove.dfm <- function(x, pattern = NULL, ...) {
-    x <- as.dfm(x)
-    dfm_select(x, pattern, selection = "remove", ...)
+dfm_keep <- function(x, ...) {
+    if ("selection" %in% names(list(...))) {
+        stop("dfm_keep cannot include selection argument")
+    }
+    dfm_select(x, ..., selection = "keep")
 }