+++ R/tokens_group.R
@@ -21,7 +21,7 @@ tokens_group <- function(x, groups = NULL) {
         x <- list(unlist(unclass(x), use.names = FALSE))
         names(x) <- as.character(groups[1])
     }
-    x <- structure(x, class = c('tokens', 'tokenizedTexts'))
+    x <- structure(x, class = 'tokens')
     docvars(x) <- data.frame(row.names = docnames(x))
     attributes(x, FALSE) <- attrs
     return(x)